feat: enhance report preview and PDF export functionality
- Extract report preview logic into reusable ReportPreviewPages component
- Create usePdfExport hook for consistent PDF generation across components
- Add PDF export button to order detail view alongside refresh button
- Change 'Ver reporte' to 'Editar reporte' in order detail
- Fix sidebar height issue on small screens with improved CSS
- Maintain PDF export functionality in report editor
- Ensure consistent preview rendering between editor and detail views
