+++ src/components/ui/sidebar_menu.tsx
@@ -84,9 +84,25 @@ const SidebarCeluma: React.FC<SidebarCelumaProps> = ({selectedKey = "/home", onN
     return (
         <>
             <style>{`
-                body, html { margin: 0; padding: 0; }
-                .ant-layout { margin: 0; padding: 0; }
-                .ant-layout-sider { margin: 0; padding: 0; }
+                body, html { 
+                    margin: 0; 
+                    padding: 0; 
+                    height: 100%;
+                    min-height: 100vh;
+                }
+                .ant-layout { 
+                    margin: 0; 
+                    padding: 0; 
+                    min-height: 100vh;
+                }
+                .ant-layout-sider { 
+                    margin: 0; 
+                    padding: 0; 
+                    min-height: 100vh;
+                }
+                #root {
+                    min-height: 100vh;
+                }
             `}</style>
             <Sider
                 width = {260}
@@ -163,13 +179,15 @@ const styles: Record<string, React.CSSProperties> = {
         paddingLeft: 0,
         paddingRight: 0,
         paddingTop: 0,
+        minHeight: "100vh",
         height: "100vh",
         position: "sticky",
         top: 0,
         overflow: "hidden",
     },
     inner: {
-        height: "100%",
+        height: "100vh",
+        minHeight: "100vh",
         display: "flex",
         flexDirection: "column",
         fontFamily: "Nanito, sans-serif",
